Project Library difficulty filters added, Minor changes to back to dashboard button and student profile viewer
