+++ src/common/BackToDashboardButton/BackToDashboardButton.module.css
@@ -7,7 +7,7 @@
   font-weight: 900;
   font-size: 0.9vw;
   padding: 5px 0px;
-  width: 13vw;
+  width: 11.5vw;
   box-shadow: 0 3px 4px rgba(0, 0, 0, 0.2);
   cursor: pointer;
 }
